Picoware - v1.6.0

This is an update to the MicroPython version that uses a PSRAM framebuffer (now you have an extra 100 kilobytes of RAM), adds auto-complete functionality in the `Keyboard` class, and more:

Added:
- Thread safety checks in HTTP class
- `Show on-screen keyboard?` toggle (in the Settings) and functionality to the `Keyboard` class: if set to false, the on-screen keyboard won't be shown in any apps/views that use the keyboard (users will need to press `CENTER`, or `RETURN`, to submit their response)
- `AutoComplete` C module (merged from Flipper Zero app) and implemented it into the `Keyboard` class
- float/int conversion methods in `Vector` class

Changes:
- Optimized Draw class and LCD C module
- Updated Waveshare 3.49 Touch LCD support: fixed input handling and decreased default font size
- Optimizations in `FlipWorld`, `Starfield`, `Patterns`, and `Flappy Bird` apps
- Optimized the QSPI PSRAM
- Updated the LCD driver to use a `PSRAM` frame buffer
- Updated the `run` method in the `Keyboard` class to return `False` if the user clicked `BACK`, which is a signal to return to the previous view
- Updated the `ViewManager` class to update the Keyboard color properties when foreground/background is changed
- Removed `get_response` and `set_response` from `Keyboard` class (use the `response` property instead)

Notes:
- To select the suggestion from the autocomplete, you need to hover over the text box (move the cursor up until the text box is highlighted) then click `CENTER` (which is the `RETURN` key on the PicoCalc). If you opted out of the on-screen keyboard (within the Picoware settings), then you simply click down from the text box and click `CENTER` on any of the suggestion(s). 
- This update has a small change that's somewhat breaking, so make sure to update all of the apps either from the on-device store or by copying and pasting the apps folder from `builds/MicroPython` into your SD card within the `picoware` folder.

+++ src/MicroPython/picoware/gui/toggle_list.py
@@ -47,9 +47,9 @@ class ToggleList:
         self.states = []
         self._selected_index = 0
         self.max_visible_items = 8
-        self.toggle_size = Vector(self.size.x - 20, self.size.y // 10.67)
+        self.toggle_size = Vector(self.size.x - 20, int(self.size.y // 10.67))
         self.toggle_position = Vector(10, 10)
-        self.toggle_spacing = self.size.y // self.max_visible_items
+        self.toggle_spacing = int(self.size.y // self.max_visible_items)
 
         self._callback = callback
         self._just_started = True
@@ -215,7 +215,7 @@ class ToggleList:
             original_y = toggle.position.y
 
             # Calculate new position for this visible slot
-            new_y = 10 + (visible_idx * self.toggle_spacing)
+            new_y = int(10 + (visible_idx * self.toggle_spacing))
             toggle.position.y = new_y
 
             # Draw the toggle
